change previous fix to a better one : add missing inclusion of cstdlib (because std::abs() is defined in stdlib.h !?!)
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/soma/soma-io/trunk@65604 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/somadicom/Transformation/PatientOrientation.cc
@@ -4,6 +4,7 @@
 #include <Transformation/PatientOrientation.h>
 #endif
 
+#include <cstdlib>
 #include <cmath>
 
 
@@ -177,18 +178,18 @@ soma::PatientOrientation::getDirectBoundingBox(
                                  double( boundingBox.getLowerY() ),
                                  double( boundingBox.getLowerZ() ),
                                  x, y, z );
-  outBoundingBox.setLowerX( std::abs( (double)int32_t( x + 0.5 ) ) );
-  outBoundingBox.setLowerY( std::abs( (double)int32_t( y + 0.5 ) ) );
-  outBoundingBox.setLowerZ( std::abs( (double)int32_t( z + 0.5 ) ) );
+  outBoundingBox.setLowerX( std::abs( int32_t( x + 0.5 ) ) );
+  outBoundingBox.setLowerY( std::abs( int32_t( y + 0.5 ) ) );
+  outBoundingBox.setLowerZ( std::abs( int32_t( z + 0.5 ) ) );
   outBoundingBox.setLowerT( boundingBox.getLowerT() );
 
   _flipTransformation.getDirect( double( boundingBox.getUpperX() ),
                                  double( boundingBox.getUpperY() ),
                                  double( boundingBox.getUpperZ() ),
                                  x, y, z );
-  outBoundingBox.setUpperX( std::abs( (double)int32_t( x + 0.5 ) ) );
-  outBoundingBox.setUpperY( std::abs( (double)int32_t( y + 0.5 ) ) );
-  outBoundingBox.setUpperZ( std::abs( (double)int32_t( z + 0.5 ) ) );
+  outBoundingBox.setUpperX( std::abs( int32_t( x + 0.5 ) ) );
+  outBoundingBox.setUpperY( std::abs( int32_t( y + 0.5 ) ) );
+  outBoundingBox.setUpperZ( std::abs( int32_t( z + 0.5 ) ) );
   outBoundingBox.setUpperT( boundingBox.getUpperT() );
 
   return outBoundingBox;
@@ -208,18 +209,18 @@ soma::PatientOrientation::getInverseBoundingBox(
                                   double( boundingBox.getLowerY() ),
                                   double( boundingBox.getLowerZ() ),
                                   x, y, z );
-  outBoundingBox.setLowerX( std::abs( (double)int32_t( x + 0.5 ) ) );
-  outBoundingBox.setLowerY( std::abs( (double)int32_t( y + 0.5 ) ) );
-  outBoundingBox.setLowerZ( std::abs( (double)int32_t( z + 0.5 ) ) );
+  outBoundingBox.setLowerX( std::abs( int32_t( x + 0.5 ) ) );
+  outBoundingBox.setLowerY( std::abs( int32_t( y + 0.5 ) ) );
+  outBoundingBox.setLowerZ( std::abs( int32_t( z + 0.5 ) ) );
   outBoundingBox.setLowerT( boundingBox.getLowerT() );
 
   _flipTransformation.getInverse( double( boundingBox.getUpperX() ),
                                   double( boundingBox.getUpperY() ),
                                   double( boundingBox.getUpperZ() ),
                                   x, y, z );
-  outBoundingBox.setUpperX( std::abs( (double)int32_t( x + 0.5 ) ) );
-  outBoundingBox.setUpperY( std::abs( (double)int32_t( y + 0.5 ) ) );
-  outBoundingBox.setUpperZ( std::abs( (double)int32_t( z + 0.5 ) ) );
+  outBoundingBox.setUpperX( std::abs( int32_t( x + 0.5 ) ) );
+  outBoundingBox.setUpperY( std::abs( int32_t( y + 0.5 ) ) );
+  outBoundingBox.setUpperZ( std::abs( int32_t( z + 0.5 ) ) );
   outBoundingBox.setUpperT( boundingBox.getUpperT() );
 
   return outBoundingBox;
